fuzzy_nucleo: Remove `util` dependency
