+++ crates/fuzzy_nucleo/Cargo.toml
@@ -16,12 +16,12 @@ doctest = false
 fuzzy.workspace = true
 nucleo.workspace = true
 gpui.workspace = true
-util.workspace = true
+gpui_util.workspace = true
+path.workspace = true
 
 [dev-dependencies]
 criterion.workspace = true
 gpui = { workspace = true, features = ["test-support"] }
-util = { workspace = true, features = ["test-support"] }
 
 [[bench]]
 name = "match_benchmark"
